rename folder charts
